Solved AoC2020 Day 20 Part 2

+++ aoc2020/aoc2020-kotlin/src/test/kotlin/de/havox_design/aoc2020/day20/Day20Test.kt
@@ -15,7 +15,7 @@ class Day20Test {
 
     @ParameterizedTest
     @MethodSource("getDataForTestProcessPart2")
-    fun testProcessPart2(filename: String, expectedResult: Long) =
+    fun testProcessPart2(filename: String, expectedResult: Int) =
         JurassicJigsaw(filename).processPart2().shouldBe(expectedResult)
 
     companion object {
@@ -28,7 +28,7 @@ class Day20Test {
         @JvmStatic
         private fun getDataForTestProcessPart2(): Stream<Arguments> =
             Stream.of(
-                Arguments.of("de/havox_design/aoc2020/day20/day20sample.txt", 273L)
+                Arguments.of("de/havox_design/aoc2020/day20/day20sample.txt", 273)
             )
     }
 }
